[#165028616] refactoring code and file name create a type CreatedMessageWithContentAndDueDate
