+++ ts/components/messages/MessageAgenda.tsx
@@ -11,7 +11,7 @@ import {
 
 import I18n from "../../i18n";
 import customVariables from "../../theme/variables";
-import { MessageWithContentAndDueDatePO } from "../../types/MessageWithContentAndDueDatePO";
+import { CreatedMessageWithContentAndDueDate } from "../../types/CreatedMessageWithContentAndDueDate";
 import H5 from "../ui/H5";
 import MessageAgendaItem from "./MessageAgendaItem";
 
@@ -39,12 +39,12 @@ const styles = StyleSheet.create({
   }
 });
 
-const keyExtractor = (_: MessageWithContentAndDueDatePO) => _.id;
+const keyExtractor = (_: CreatedMessageWithContentAndDueDate) => _.id;
 
 const ItemSeparatorComponent = () => <View style={styles.itemSeparator} />;
 
 export type MessageAgendaSection = SectionListData<
-  MessageWithContentAndDueDatePO
+  CreatedMessageWithContentAndDueDate
 >;
 
 type Props = {
@@ -106,7 +106,7 @@ class MessageAgenda extends React.PureComponent<Props> {
   };
 
   private renderItem: SectionListRenderItem<
-    MessageWithContentAndDueDatePO
+    CreatedMessageWithContentAndDueDate
   > = info => {
     const message = info.item;
     return (
